#302 refactor: 기능 및 책임에 맞게 컨트롤러 분리
