+++ src/main/java/UMC_7th/Closit/domain/user/service/UserQueryServiceImpl.java
@@ -34,19 +34,18 @@ public class UserQueryServiceImpl implements UserQueryService {
     private final PostRepository postRepository;
     private final SecurityUtil securityUtil;
     private final BlockRepository userBlockRepository;
+    private final UserUtil userUtil;
 
     @Override
     public Slice<Highlight> getHighlightList(String clositId, Pageable pageable) {
-        User user = userRepository.findByClositId(clositId)
-                .orElseThrow(() -> new UserHandler(ErrorStatus.USER_NOT_FOUND));
+        User user = userUtil.getUserByClositIdOrThrow(clositId);
 
         return highlightRepository.findAllByUser(user, pageable);
     }
 
     @Override
     public Slice<User> getFollowerList(String clositId, Pageable pageable) {
-        User user = userRepository.findByClositId(clositId)
-                .orElseThrow(() -> new UserHandler(ErrorStatus.USER_NOT_FOUND));
+        User user = userUtil.getUserByClositIdOrThrow(clositId);
 
         Slice<Follow> followers = followRepository.findByReceiver(user, pageable);
         return followers.map(Follow::getSender);
@@ -54,8 +53,7 @@ public class UserQueryServiceImpl implements UserQueryService {
 
     @Override
     public Slice<User> getFollowingList(String clositId, Pageable pageable) {
-        User user = userRepository.findByClositId(clositId)
-                .orElseThrow(() -> new UserHandler(ErrorStatus.USER_NOT_FOUND));
+        User user = userUtil.getUserByClositIdOrThrow(clositId);
 
         Slice<Follow> followings = followRepository.findBySender(user, pageable);
         return followings.map(Follow::getReceiver);
@@ -63,8 +61,7 @@ public class UserQueryServiceImpl implements UserQueryService {
 
     @Override
     public UserResponseDTO.UpdateUserInfoDTO getUserInfo(String clositId) {
-        User user = userRepository.findByClositId(clositId)
-                .orElseThrow(() -> new UserHandler(ErrorStatus.USER_NOT_FOUND));
+        User user = userUtil.getUserByClositIdOrThrow(clositId);
 
         long followerCount = followRepository.countByReceiver(user);
         long followingCount = followRepository.countBySender(user);
@@ -74,7 +71,6 @@ public class UserQueryServiceImpl implements UserQueryService {
 
     @Override
     public boolean isMissionDone() {
-        // 현재 로그인된 사용자 정보 가져오기
         User user = securityUtil.getCurrentUser();
 
         LocalDateTime startOfDay = LocalDateTime.now().toLocalDate().atStartOfDay();
@@ -87,7 +83,6 @@ public class UserQueryServiceImpl implements UserQueryService {
     @Override
     public Slice<Post> getRecentPostList(String clositId, Integer page) { // 특정 사용자의 최근 게시글 조회
         Pageable pageable = PageRequest.of(page, 10);
-
         LocalDateTime week = LocalDateTime.now().minusWeeks(1);
 
         Slice<Post> recentPostList = postRepository.findRecentPostList(clositId, week, pageable);
@@ -98,22 +93,20 @@ public class UserQueryServiceImpl implements UserQueryService {
     @Override
     public Slice<String> getBlockedUserList(Pageable pageable) {
         User currentUser = securityUtil.getCurrentUser();
+
         return userBlockRepository.findBlockedUsersByBlocker(currentUser.getClositId(), pageable);
     }
 
     @Override
     public UserResponseDTO.IsBlockedDTO isBlockedBy(String targetClositId) {
+        userUtil.getUserByClositIdOrThrow(targetClositId);
 
-        User targetUser = userRepository.findByClositId(targetClositId)
-                .orElseThrow(() -> new UserHandler(ErrorStatus.USER_NOT_FOUND));
+        String currentClositId = securityUtil.getCurrentUser().getClositId();
+        boolean isBlocked = userBlockRepository.existsByBlockerIdAndBlockedId(targetClositId, currentClositId);
 
-        User currentUser = securityUtil.getCurrentUser();
-
-        UserResponseDTO.IsBlockedDTO isblockedDTO = UserResponseDTO.IsBlockedDTO.builder()
-                .isBlocked(userBlockRepository.existsByBlockerIdAndBlockedId(targetClositId, currentUser.getClositId()))
+        return UserResponseDTO.IsBlockedDTO.builder()
+                .isBlocked(isBlocked)
                 .build();
-
-        return isblockedDTO;
     }
 
     @Override
